+++ golem-worker-service-base/src/parser/expr/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod expr_parser;
 mod flags;
 mod let_statement;
 mod record;
+mod selection;
 mod sequence;
 mod tuple;
 mod util;
